Added SPF and IFCXML format
Moved some stuff around from IfcDoc.Program.

Worked on Compiler a bit to give it functionalities to build IfcProjectLibrary.

Compiler would require a re-work as there is now a lot of static stuff in it.
