+++ Formats/IfcDoc.Formats/IfcDoc.Formats.csproj
@@ -11,6 +11,9 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\IfcDoc.Schemas\IfcDoc.Schemas\IfcDoc.Schemas.csproj" />
+    <ProjectReference Include="..\..\IfcKit\formats\BuildingSmart.Serialization\BuildingSmart.Serialization.csproj" />
+    <ProjectReference Include="..\..\IfcKit\formats\spf\BuildingSmart.Serialization.Step.csproj" />
+    <ProjectReference Include="..\..\IfcKit\formats\xml\BuildingSmart.Serialization.Xml.csproj" />
   </ItemGroup>
 
   <ItemGroup>
