+++ Program.cs
@@ -1023,21 +1023,7 @@ namespace IfcDoc
 			return docSchema;
 		}
 
-		internal static void ExportIfcDefinition(
-			IfcPropertyTemplateDefinition ifcDefinition,
-			DocObject docObject)
-		{
-			ifcDefinition.GlobalId = new IfcGloballyUniqueId(GlobalId.Format(docObject.Uuid));
-			ifcDefinition.Name = new IfcLabel(docObject.Name);
-			ifcDefinition.Description = new IfcText(docObject.Documentation);
 
-			foreach (DocLocalization docLoc in docObject.Localization)
-			{
-				IfcLibraryReference ifcLib = new IfcLibraryReference(new IfcURIReference(docLoc.URL), null, new IfcLabel(docLoc.Name), new IfcText(docLoc.Documentation), new IfcLanguageId(docLoc.Locale), null);
-				IfcRelAssociatesLibrary ifcRal = new IfcRelAssociatesLibrary(NewGuid(), null, null, null, new IfcDefinitionSelect[] { ifcDefinition }, ifcLib);
-				ifcDefinition.HasAssociations.Add(ifcRal);
-			}
-		}
 
 		internal static void SetObject(DocObject obj, IfcRoot root)
 		{
@@ -1407,262 +1393,9 @@ namespace IfcDoc
 			string pathLog = System.IO.Path.Combine(System.IO.Path.GetDirectoryName(filePath), System.IO.Path.GetFileNameWithoutExtension(filePath)+ ".txt");
 			System.IO.File.WriteAllText(pathLog, string.Join("\r\n", changes));
 			System.Diagnostics.Process.Start(pathLog);
-		}
+		}	
 
-		
-		internal static void ExportIfc(IfcProjectLibrary ifcProjectLibrary, DocProject docProject, Dictionary<DocObject, bool> included)
-		{
-			//List<IfcDefinitionSelect> propertyEnumerations = new List<IfcDefinitionSelect>();  IfcPropertyEnumeration doesn't inherit from IfcDefinitionSelect
-			List<IfcDefinitionSelect> simpleProperties = new List<IfcDefinitionSelect>();
-			List<IfcDefinitionSelect> complexProperties = new List<IfcDefinitionSelect>();
-			List<IfcDefinitionSelect> propertySets = new List<IfcDefinitionSelect>();
-			List<IfcDefinitionSelect> quantities = new List<IfcDefinitionSelect>();
-			List<IfcDefinitionSelect> quantitySets = new List<IfcDefinitionSelect>();
 
-			Dictionary<string, DocPropertyEnumeration> enumerations = new Dictionary<string, DocPropertyEnumeration>();
-			Dictionary<string, DocProperty> simplePropertyDictionary = new Dictionary<string, DocProperty>();
-			Dictionary<string, DocProperty> complexPropertyDictionary = new Dictionary<string, DocProperty>();
-			List<DocQuantity> simpleQuantityList = new List<DocQuantity>();
-			List<DocPropertySet> propertySetList = new List<DocPropertySet>();
-			List<DocQuantitySet> quantitySetList = new List<DocQuantitySet>();
-
-			foreach (DocSection docSection in docProject.Sections)
-			{
-				foreach (DocSchema docSchema in docSection.Schemas)
-				{
-					foreach (DocPropertySet docPset in docSchema.PropertySets)
-					{
-						if (included == null || included.ContainsKey(docPset))
-						{
-							propertySetList.Add(docPset);
-							foreach (DocProperty docProp in docPset.Properties)
-								processProperty(docProp, simplePropertyDictionary, complexPropertyDictionary, enumerations);
-						}
-					}
-					foreach (DocQuantitySet docQuantitySet in docSchema.QuantitySets)
-					{
-						if (included == null || included.ContainsKey(docQuantitySet))
-						{
-							quantitySetList.Add(docQuantitySet);
-							foreach (DocQuantity docQuantity in docQuantitySet.Quantities)
-								simpleQuantityList.Add(docQuantity);
-						}
-					}
-				}
-			}
-			DocObject.ComparerDocObject comparer = new DocObject.ComparerDocObject();
-			Dictionary<string, IfcPropertyTemplate> propertyTemplates = new Dictionary<string, IfcPropertyTemplate>();
-			Dictionary<string, IfcPropertyEnumeration> dictionaryEnumerations = new Dictionary<string, IfcPropertyEnumeration>();
-			foreach (DocPropertyEnumeration docEnumeration in enumerations.Values.OrderBy(x=>x,comparer))
-			{
-				// cache enumerators
-				try
-				{
-					IfcPropertyEnumeration ifcProp =  new IfcPropertyEnumeration(new IfcLabel(docEnumeration.Name), new IfcValue[] { }, null);
-					dictionaryEnumerations[docEnumeration.Name] = ifcProp; //
-				//	propertyEnumerations.Add(ifcProp);
-					foreach (DocPropertyConstant docConst in docEnumeration.Constants)
-					{
-						IfcLabel label = new IfcLabel(docConst.Name.ToUpper());
-						ifcProp.EnumerationValues.Add(label);
-
-						foreach (DocLocalization docLoc in docConst.Localization)
-						{
-							IfcLibraryReference ifcLib = new IfcLibraryReference(null, null, null, null, null, null);
-							if (!string.IsNullOrEmpty(docLoc.URL))
-								ifcLib.Location = new IfcURIReference(docLoc.URL);
-							if (!string.IsNullOrEmpty(docConst.Name))
-							{
-								ifcLib.Identification = new IfcIdentifier(docConst.Name);
-								ifcLib.Name = new IfcLabel(docLoc.Name);
-							}
-							if (!string.IsNullOrEmpty(docLoc.Documentation))
-								ifcLib.Description = new IfcText(docLoc.Documentation);
-							if (!string.IsNullOrEmpty(docLoc.Locale))
-								ifcLib.Language = new IfcLanguageId(docLoc.Locale);
-
-							if (docLoc.Documentation == null && docLoc.Locale == "en")
-							{
-								ifcLib.Description = new IfcText(docConst.Documentation);
-							}
-
-							//IfcRelAssociatesLibrary ifcRal = new IfcRelAssociatesLibrary(NewGuid(), null, null, null, new IfcDefinitionSelect[] { ifcProp }, ifcLib);
-							//ifcProp.HasAssociations.Add(ifcRal);
-						}
-
-					}
-				}
-				catch
-				{
-				}
-			}
-			foreach(DocProperty docProperty in simplePropertyDictionary.Values.OrderBy(x=> x,comparer))
-			{
-				IfcPropertyTemplate ifcProperty = convertProperty(docProperty, dictionaryEnumerations, propertyTemplates);
-				simpleProperties.Add(ifcProperty);
-			}
-
-			foreach (DocProperty docProperty in complexPropertyDictionary.Values.OrderBy(x => x, comparer))
-			{
-				IfcPropertyTemplate ifcProperty = convertProperty(docProperty, dictionaryEnumerations, propertyTemplates);
-				complexProperties.Add(ifcProperty);
-			}
-			
-
-			foreach (DocPropertySet docPset in propertySetList.OrderBy(x => x, comparer))
-			{
-				IfcPropertySetTemplate ifcPset = new IfcPropertySetTemplate(Program.NewGuid(), null, null, null, null, null, new IfcPropertyTemplate[] { });
-
-				ExportIfcDefinition(ifcPset, docPset);
-
-				switch (docPset.PropertySetType)
-				{
-					case "PSET_TYPEDRIVENOVERRIDE":
-						ifcPset.TemplateType = IfcPropertySetTemplateTypeEnum.PSET_TYPEDRIVENOVERRIDE;
-						break;
-
-					case "PSET_OCCURRENCEDRIVEN":
-						ifcPset.TemplateType = IfcPropertySetTemplateTypeEnum.PSET_OCCURRENCEDRIVEN;
-						break;
-
-					case "PSET_PERFORMANCEDRIVEN":
-						ifcPset.TemplateType = IfcPropertySetTemplateTypeEnum.PSET_PERFORMANCEDRIVEN;
-						break;
-				}
-
-				ifcPset.ApplicableEntity = new IfcIdentifier(docPset.ApplicableType);
-				propertySets.Add(ifcPset);
-
-				foreach (DocProperty docProp in docPset.Properties)
-					ifcPset.HasPropertyTemplates.Add(propertyTemplates[docProp.UniqueId]);	
-			}
-
-			foreach (DocQuantity docQuantity in simpleQuantityList.OrderBy(x => x, comparer))
-			{
-				IfcSimplePropertyTemplate ifcProperty = new IfcSimplePropertyTemplate(Program.NewGuid(), null, null, null, null, null, null, null, null, null, null, null);
-				ExportIfcDefinition(ifcProperty, docQuantity);
-				ifcProperty.TemplateType = (IfcSimplePropertyTemplateTypeEnum)Enum.Parse(typeof(IfcSimplePropertyTemplateTypeEnum), docQuantity.QuantityType.ToString());
-				propertyTemplates[docQuantity.UniqueId] = ifcProperty;
-				quantities.Add(ifcProperty);
-			}
-			foreach (DocQuantitySet docQuantitySet in quantitySetList.OrderBy(x=>x,comparer))
-			{
-				IfcPropertySetTemplate ifcPset = new IfcPropertySetTemplate(Program.NewGuid(), null, null, null, null, null, new IfcPropertyTemplate[] { });
-
-				ExportIfcDefinition(ifcPset, docQuantitySet);
-
-				ifcPset.TemplateType = IfcPropertySetTemplateTypeEnum.QTO_OCCURRENCEDRIVEN;
-				ifcPset.ApplicableEntity = new IfcIdentifier(docQuantitySet.ApplicableType);
-				quantitySets.Add(ifcPset);
-
-				foreach (DocQuantity docQuantity in docQuantitySet.Quantities)
-					ifcPset.HasPropertyTemplates.Add(propertyTemplates[docQuantity.UniqueId]);
-			}
-
-			//if (propertyEnumerations.Count > 0)
-			//	ifcProjectLibrary.Declares.Add(new IfcRelDeclares(new IfcGloballyUniqueId(), null, new IfcLabel("PropertyEnumerations"), null, ifcProjectLibrary, propertyEnumerations.ToArray()));
-			if (simpleProperties.Count > 0)
-				ifcProjectLibrary.Declares.Add(new IfcRelDeclares(new IfcGloballyUniqueId("3RyLNYXkX479wspBr_nf6n"), null, new IfcLabel("SimpleProperties"), null, ifcProjectLibrary, simpleProperties.ToArray()));
-			if(complexProperties.Count > 0)
-				ifcProjectLibrary.Declares.Add(new IfcRelDeclares(new IfcGloballyUniqueId("0dJzky7sb6txqIsyVpv7iW"), null, new IfcLabel("ComplexProperties"), null, ifcProjectLibrary, complexProperties.ToArray()));
-			if (propertySets.Count > 0)
-				ifcProjectLibrary.Declares.Add(new IfcRelDeclares(new IfcGloballyUniqueId("3M4U3yWxr3_gsGvVZ3qXJS"), null, new IfcLabel("PropertySets"), null, ifcProjectLibrary, propertySets.ToArray()));
-			if (quantities.Count > 0)
-				ifcProjectLibrary.Declares.Add(new IfcRelDeclares(new IfcGloballyUniqueId("1$c0OOaRDF_OBDoEs$xzcD"), null, new IfcLabel("Quantities"), null, ifcProjectLibrary, quantities.ToArray()));
-			if (quantitySets.Count > 0)
-				ifcProjectLibrary.Declares.Add(new IfcRelDeclares(new IfcGloballyUniqueId("1CpAEUB6v1Fhb57tduNMhL"), null, new IfcLabel("QuantitySets"), null, ifcProjectLibrary, quantitySets.ToArray()));
-		}
-		private static IfcPropertyTemplate convertProperty(DocProperty property, Dictionary<string, IfcPropertyEnumeration> mapEnums, Dictionary<string,IfcPropertyTemplate> properties)
-		{
-			if (property.PropertyType == DocPropertyTemplateTypeEnum.COMPLEX)
-			{
-				IfcComplexPropertyTemplate ifcProp = new IfcComplexPropertyTemplate(Program.NewGuid(), null, null, null, null, null, new IfcPropertyTemplate[0]);
-				ExportIfcDefinition(ifcProp, property);
-				ifcProp.TemplateType = IfcComplexPropertyTemplateTypeEnum.P_COMPLEX;
-				foreach (DocProperty element in property.Elements)
-				{
-					IfcPropertyTemplate propertyTemplate = null;
-					if (properties.TryGetValue(element.UniqueId, out propertyTemplate))
-						ifcProp.HasPropertyTemplates.Add(propertyTemplate);
-					else
-					{
-						ifcProp.HasPropertyTemplates.Add(convertProperty(element, mapEnums, properties));
-					}
-				}
-				return properties[property.UniqueId] = ifcProp;
-			}
-			return properties[property.UniqueId] = ExportIfcPropertyTemplate(property, mapEnums);
-		}
-		private static void processProperty(DocProperty property, Dictionary<string, DocProperty> simpleProperties, Dictionary<string, DocProperty> complexProperties, Dictionary<string, DocPropertyEnumeration> enumerations)
-		{
-			if (property.PropertyType == DocPropertyTemplateTypeEnum.COMPLEX)
-			{
-				if (!complexProperties.ContainsKey(property.UniqueId))
-				{
-					complexProperties[property.UniqueId] = property;
-					foreach (DocProperty element in property.Elements)
-						processProperty(element, simpleProperties, complexProperties, enumerations);
-				}
-			}
-			else if (!simpleProperties.ContainsKey(property.UniqueId))
-			{
-				simpleProperties[property.UniqueId] = property;
-				if (property.PropertyType == DocPropertyTemplateTypeEnum.P_ENUMERATEDVALUE)
-				{
-					if (property.Enumeration != null)
-					{
-						if (!enumerations.ContainsKey(property.Enumeration.Name))
-							enumerations[property.Enumeration.Name] = property.Enumeration;
-					}
-					else
-						Debug.WriteLine("XXX Missing Enumeration " + property.Name);
-				}
-			}
-		}
-		
-
-		private static IfcPropertyTemplate ExportIfcPropertyTemplate( DocProperty docProp, Dictionary<string, IfcPropertyEnumeration> mapEnums)
-		{
-			if (docProp.PropertyType == DocPropertyTemplateTypeEnum.COMPLEX)
-			{
-				IfcComplexPropertyTemplate ifcProp = new IfcComplexPropertyTemplate(Program.NewGuid(), null, null, null, null, null, new IfcPropertyTemplate[] { });
-
-				ExportIfcDefinition(ifcProp, docProp);
-
-				ifcProp.TemplateType = IfcComplexPropertyTemplateTypeEnum.P_COMPLEX;
-				ifcProp.UsageName = new IfcLabel(docProp.PrimaryDataType);
-
-				foreach (DocProperty docSubProp in docProp.Elements)
-				{
-					IfcPropertyTemplate ifcSub = ExportIfcPropertyTemplate(docSubProp, mapEnums);
-					ifcProp.HasPropertyTemplates.Add(ifcSub);
-				}
-
-				return ifcProp;
-			}
-			else
-			{
-				IfcSimplePropertyTemplate ifcProp = new IfcSimplePropertyTemplate(Program.NewGuid(), null, null, null, null, null, null, null, null, null, null, null);
-
-				ExportIfcDefinition(ifcProp, docProp);
-
-				ifcProp.TemplateType = (IfcSimplePropertyTemplateTypeEnum)Enum.Parse(typeof(IfcSimplePropertyTemplateTypeEnum), docProp.PropertyType.ToString());
-				ifcProp.PrimaryMeasureType = new IfcLabel(docProp.PrimaryDataType);
-				if(!string.IsNullOrEmpty(docProp.SecondaryDataType))
-					ifcProp.SecondaryMeasureType = new IfcLabel(docProp.SecondaryDataType);
-
-				// enumerations
-				if (ifcProp.TemplateType == IfcSimplePropertyTemplateTypeEnum.P_ENUMERATEDVALUE)
-				{
-					if(docProp.Enumeration != null)
-						ifcProp.Enumerators = mapEnums[docProp.Enumeration.Name];
-					else
-						System.Diagnostics.Debug.WriteLine("XXX Missing Enumeration " + docProp.Name);
-				}
-
-				return ifcProp;
-			}
-		}
 
 		public static DocPropertySet ImportPsd(PropertySetDef psd, DocProject docProject)
 		{
